+++ vulkano/src/pipeline/shader.rs
@@ -166,7 +166,7 @@ impl ShaderModule {
 unsafe impl VulkanObject for ShaderModule {
     type Object = vk::ShaderModule;
 
-    const TYPE: vk::DebugReportObjectTypeEXT = vk::DEBUG_REPORT_OBJECT_TYPE_SHADER_MODULE_EXT;
+    const TYPE: vk::ObjectType= vk::OBJECT_TYPE_SHADER_MODULE;
 
     #[inline]
     fn internal_object(&self) -> vk::ShaderModule {