+++ vulkano/src/swapchain/display.rs
@@ -310,7 +310,7 @@ impl Display {
 unsafe impl VulkanObject for Display {
     type Object = vk::DisplayKHR;
 
-    const TYPE: vk::DebugReportObjectTypeEXT = vk::DEBUG_REPORT_OBJECT_TYPE_DISPLAY_KHR_EXT;
+    const TYPE: vk::ObjectType= vk::OBJECT_TYPE_DISPLAY_KHR;
 
     #[inline]
     fn internal_object(&self) -> vk::DisplayKHR {
@@ -380,7 +380,7 @@ impl DisplayMode {
 unsafe impl VulkanObject for DisplayMode {
     type Object = vk::DisplayModeKHR;
 
-    const TYPE: vk::DebugReportObjectTypeEXT = vk::DEBUG_REPORT_OBJECT_TYPE_DISPLAY_MODE_KHR_EXT;
+    const TYPE: vk::ObjectType= vk::OBJECT_TYPE_DISPLAY_MODE_KHR;
 
     #[inline]
     fn internal_object(&self) -> vk::DisplayModeKHR {