+++ vulkano/src/sync/event.rs
@@ -164,7 +164,7 @@ unsafe impl DeviceOwned for Event {
 unsafe impl VulkanObject for Event {
     type Object = vk::Event;
 
-    const TYPE: vk::DebugReportObjectTypeEXT = vk::DEBUG_REPORT_OBJECT_TYPE_EVENT_EXT;
+    const TYPE: vk::ObjectType= vk::OBJECT_TYPE_EVENT;
 
     #[inline]
     fn internal_object(&self) -> vk::Event {