+++ vulkano/src/sync/fence.rs
@@ -307,7 +307,7 @@ unsafe impl<D> VulkanObject for Fence<D>
 {
     type Object = vk::Fence;
 
-    const TYPE: vk::DebugReportObjectTypeEXT = vk::DEBUG_REPORT_OBJECT_TYPE_FENCE_EXT;
+    const TYPE: vk::ObjectType= vk::OBJECT_TYPE_FENCE;
 
     #[inline]
     fn internal_object(&self) -> vk::Fence {